requirements.txt

aiogram==2.25.1
APScheduler==3.10.4
python-dotenv==1.0.1
aiosqlite==0.20.0
loguru==0.7.2
